(Refact) seperate pdf uploading API code
