+++ src/features/Member/resume/upload.tsx
@@ -1,15 +1,12 @@
 import { Button } from '@/components/ui/button';
 
 import fileUpload from '@/features/Member/resume/image/file-arrow-up-light.svg';
-import { handlePresignedUrl } from '@/features/Member/profile/util/handlePresignedUrl';
-import { useAuthStore } from '@/features/Member/auth/store/auth';
-import { useResumeStore } from '@/features/Member/resume/store/resumeStore';
 
-import { safePost } from '@/lib/request';
+import { useAuthStore } from '@/features/Member/auth/store/auth';
+import { uploadPdf } from './util/uploadPdf';
+
 import { useForm, Controller } from 'react-hook-form';
 import { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import React from 'react';
 
 export default function Upload() {
   const token = useAuthStore((state) => state.token);
@@ -35,43 +32,6 @@ export default function Upload() {
   }, [watch('resume')]);
 
   const [fileUrl, setFileUrl] = useState<string | null>(null);
-  const navigate = useNavigate();
-  // const submitForm = () => {
-  const submitForm = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    const input = (e.target as HTMLFormElement).querySelector<HTMLInputElement>('#resume-upload');
-    const file = input?.files?.[0] || null;
-    if (file != null) {
-      const objectUrl = await handlePresignedUrl({
-        file,
-        token,
-        type: 'resume',
-        uploadType: 'RESUME',
-      });
-
-      if (!token) return;
-      try {
-        const res = await safePost(
-          '/api/v1/members/resume/complete-upload',
-          {
-            objectUrl: objectUrl,
-          },
-          {
-            headers: { Authorization: `Bearer ${token}` },
-          },
-        );
-        if (res.httpStatusCode === 200) {
-          useResumeStore.getState().setResume(res.data);
-          window.location.href = '/resume/form';
-        }
-      } catch (err: any) {
-        alert(err);
-      }
-    }
-
-    navigate('/resume/form');
-  };
 
   return (
     <>
@@ -83,7 +43,7 @@ export default function Upload() {
             <p>이력서를 업로드하면, 더 정밀한 분석 결과를 받아볼 수 있어요.</p>
           </div>
         </div>
-        <form onSubmit={submitForm}>
+        <form onSubmit={(e) => uploadPdf(e, token)}>
           <div className="flex flex-col w-full items-center gap-2">
             <label
               htmlFor="resume-upload"
